+++ src/main/java/http/Cookies.java
@@ -1,4 +1,4 @@
-package domain;
+package http;
 
 import java.util.Arrays;
 import java.util.HashMap;
