+++ src/main/java/http/response/ResponseBody.java
@@ -1,4 +1,4 @@
-package domain;
+package http.response;
 
 import java.nio.charset.StandardCharsets;
