+++ src/main/java/http/response/ResponseHeaders.java
@@ -1,4 +1,4 @@
-package domain;
+package http.response;
 
 import java.util.HashMap;
 import java.util.Map;
