+++ src/main/java/webserver/RequestHandler.java
@@ -4,14 +4,17 @@ import java.io.*;
 import java.net.Socket;
 import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import controller.Dispatcher;
 import controller.LoginController;
 import controller.UserController;
-import domain.*;
+import http.*;
+import http.request.Request;
+import http.response.Response;
+import http.response.ResponseBody;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.FileIoUtils;
